runtime: add stream flags to io_signature
This adds a new parameter to the io_signature class to allow specifying
flags for each input or output port.  The initial use case for this
capability is to allow a block to indicate that it needs to own and supply
the stream buffer memory for one or more of its outputs, or that upstream
connected block outputs need to query this block to provide the stream
buffer memory for the connected input/output.